@@ -1,6 +1,6 @@
-import * as React from 'react';
-import { Component } from 'react';
-import Login from '../components/Login';
+import * as React from "react";
+import { Component } from "react";
+import Login from "../components/Login";
 
 export default class LoginPage extends Component {
   render() {
