@@ -1,25 +1,25 @@
-import * as React from 'react';
-import { render } from 'react-dom';
-import { AppContainer } from 'react-hot-loader';
-import Root from './containers/Root';
-import './app.global.css';
+import * as React from "react";
+import { render } from "react-dom";
+import { AppContainer } from "react-hot-loader";
+import Root from "./containers/Root";
+import "./app.global.css";
 
 render(
   <AppContainer>
     <Root />
   </AppContainer>,
-  document.getElementById('root')
+  document.getElementById("root")
 );
 
 if ((module as any).hot) {
-  (module as any).hot.accept('./containers/Root', () => {
+  (module as any).hot.accept("./containers/Root", () => {
     // eslint-disable-next-line global-require
-    const NextRoot = require('./containers/Root').default;
+    const NextRoot = require("./containers/Root").default;
     render(
       <AppContainer>
         <NextRoot />
       </AppContainer>,
-      document.getElementById('root')
+      document.getElementById("root")
     );
   });
 }
