@@ -4,7 +4,7 @@ import {
   shell,
   BrowserWindow,
   MenuItemConstructorOptions
-} from 'electron';
+} from "electron";
 
 export default class MenuBuilder {
   mainWindow: BrowserWindow;
@@ -15,14 +15,14 @@ export default class MenuBuilder {
 
   buildMenu() {
     if (
-      process.env.NODE_ENV === 'development' ||
-      process.env.DEBUG_PROD === 'true'
+      process.env.NODE_ENV === "development" ||
+      process.env.DEBUG_PROD === "true"
     ) {
       this.setupDevelopmentEnvironment();
     }
 
     const template =
-      process.platform === 'darwin'
+      process.platform === "darwin"
         ? this.buildDarwinTemplate()
         : this.buildDefaultTemplate();
 
@@ -34,12 +34,12 @@ export default class MenuBuilder {
 
   setupDevelopmentEnvironment() {
     this.mainWindow.webContents.toggleDevTools();
-    this.mainWindow.webContents.on('context-menu', (e, props) => {
+    this.mainWindow.webContents.on("context-menu", (e, props) => {
       const { x, y } = props;
 
       Menu.buildFromTemplate([
         {
-          label: 'Inspect element',
+          label: "Inspect element",
           click: () => {
             this.mainWindow.webContents.inspectElement(x, y);
           }
@@ -50,30 +50,30 @@ export default class MenuBuilder {
 
   buildDarwinTemplate(): MenuItemConstructorOptions[] {
     const subMenuAbout = {
-      label: 'Electron',
+      label: "Electron",
       submenu: [
         {
-          label: 'About ElectronReact',
-          selector: 'orderFrontStandardAboutPanel:'
+          label: "About ElectronReact",
+          selector: "orderFrontStandardAboutPanel:"
         },
-        { type: 'separator' },
-        { label: 'Services', submenu: [] },
-        { type: 'separator' },
+        { type: "separator" },
+        { label: "Services", submenu: [] },
+        { type: "separator" },
         {
-          label: 'Hide ElectronReact',
-          accelerator: 'Command+H',
-          selector: 'hide:'
+          label: "Hide ElectronReact",
+          accelerator: "Command+H",
+          selector: "hide:"
         },
         {
-          label: 'Hide Others',
-          accelerator: 'Command+Shift+H',
-          selector: 'hideOtherApplications:'
+          label: "Hide Others",
+          accelerator: "Command+Shift+H",
+          selector: "hideOtherApplications:"
         },
-        { label: 'Show All', selector: 'unhideAllApplications:' },
-        { type: 'separator' },
+        { label: "Show All", selector: "unhideAllApplications:" },
+        { type: "separator" },
         {
-          label: 'Quit',
-          accelerator: 'Command+Q',
+          label: "Quit",
+          accelerator: "Command+Q",
           click: () => {
             app.quit();
           }
@@ -81,41 +81,41 @@ export default class MenuBuilder {
       ]
     };
     const subMenuEdit = {
-      label: 'Edit',
+      label: "Edit",
       submenu: [
-        { label: 'Undo', accelerator: 'Command+Z', selector: 'undo:' },
-        { label: 'Redo', accelerator: 'Shift+Command+Z', selector: 'redo:' },
-        { type: 'separator' },
-        { label: 'Cut', accelerator: 'Command+X', selector: 'cut:' },
-        { label: 'Copy', accelerator: 'Command+C', selector: 'copy:' },
-        { label: 'Paste', accelerator: 'Command+V', selector: 'paste:' },
+        { label: "Undo", accelerator: "Command+Z", selector: "undo:" },
+        { label: "Redo", accelerator: "Shift+Command+Z", selector: "redo:" },
+        { type: "separator" },
+        { label: "Cut", accelerator: "Command+X", selector: "cut:" },
+        { label: "Copy", accelerator: "Command+C", selector: "copy:" },
+        { label: "Paste", accelerator: "Command+V", selector: "paste:" },
         {
-          label: 'Select All',
-          accelerator: 'Command+A',
-          selector: 'selectAll:'
+          label: "Select All",
+          accelerator: "Command+A",
+          selector: "selectAll:"
         }
       ]
     };
     const subMenuViewDev = {
-      label: 'View',
+      label: "View",
       submenu: [
         {
-          label: 'Reload',
-          accelerator: 'Command+R',
+          label: "Reload",
+          accelerator: "Command+R",
           click: () => {
             this.mainWindow.webContents.reload();
           }
         },
         {
-          label: 'Toggle Full Screen',
-          accelerator: 'Ctrl+Command+F',
+          label: "Toggle Full Screen",
+          accelerator: "Ctrl+Command+F",
           click: () => {
             this.mainWindow.setFullScreen(!this.mainWindow.isFullScreen());
           }
         },
         {
-          label: 'Toggle Developer Tools',
-          accelerator: 'Alt+Command+I',
+          label: "Toggle Developer Tools",
+          accelerator: "Alt+Command+I",
           click: () => {
             this.mainWindow.webContents.toggleDevTools();
           }
@@ -123,11 +123,11 @@ export default class MenuBuilder {
       ]
     };
     const subMenuViewProd = {
-      label: 'View',
+      label: "View",
       submenu: [
         {
-          label: 'Toggle Full Screen',
-          accelerator: 'Ctrl+Command+F',
+          label: "Toggle Full Screen",
+          accelerator: "Ctrl+Command+F",
           click: () => {
             this.mainWindow.setFullScreen(!this.mainWindow.isFullScreen());
           }
@@ -135,52 +135,52 @@ export default class MenuBuilder {
       ]
     };
     const subMenuWindow = {
-      label: 'Window',
+      label: "Window",
       submenu: [
         {
-          label: 'Minimize',
-          accelerator: 'Command+M',
-          selector: 'performMiniaturize:'
+          label: "Minimize",
+          accelerator: "Command+M",
+          selector: "performMiniaturize:"
         },
-        { label: 'Close', accelerator: 'Command+W', selector: 'performClose:' },
-        { type: 'separator' },
-        { label: 'Bring All to Front', selector: 'arrangeInFront:' }
+        { label: "Close", accelerator: "Command+W", selector: "performClose:" },
+        { type: "separator" },
+        { label: "Bring All to Front", selector: "arrangeInFront:" }
       ]
     };
     const subMenuHelp = {
-      label: 'Help',
+      label: "Help",
       submenu: [
         {
-          label: 'Learn More',
+          label: "Learn More",
           click() {
-            shell.openExternal('http://electron.atom.io');
+            shell.openExternal("http://electron.atom.io");
           }
         },
         {
-          label: 'Documentation',
+          label: "Documentation",
           click() {
             shell.openExternal(
-              'https://github.com/atom/electron/tree/master/docs#readme'
+              "https://github.com/atom/electron/tree/master/docs#readme"
             );
           }
         },
         {
-          label: 'Community Discussions',
+          label: "Community Discussions",
           click() {
-            shell.openExternal('https://discuss.atom.io/c/electron');
+            shell.openExternal("https://discuss.atom.io/c/electron");
           }
         },
         {
-          label: 'Search Issues',
+          label: "Search Issues",
           click() {
-            shell.openExternal('https://github.com/atom/electron/issues');
+            shell.openExternal("https://github.com/atom/electron/issues");
           }
         }
       ]
     };
 
     const subMenuView =
-      process.env.NODE_ENV === 'development' ? subMenuViewDev : subMenuViewProd;
+      process.env.NODE_ENV === "development" ? subMenuViewDev : subMenuViewProd;
 
     return [
       subMenuAbout,
@@ -194,12 +194,12 @@ export default class MenuBuilder {
   buildDefaultTemplate(): MenuItemConstructorOptions[] {
     const templateDefault = [
       {
-        label: '&File',
+        label: "&File",
         submenu: [
-          { label: '&Open', accelerator: 'Ctrl+O' },
+          { label: "&Open", accelerator: "Ctrl+O" },
           {
-            label: '&Close',
-            accelerator: 'Ctrl+W',
+            label: "&Close",
+            accelerator: "Ctrl+W",
             click: () => {
               this.mainWindow.close();
             }
@@ -207,20 +207,20 @@ export default class MenuBuilder {
         ]
       },
       {
-        label: '&View',
+        label: "&View",
         submenu:
-          process.env.NODE_ENV === 'development'
+          process.env.NODE_ENV === "development"
             ? [
                 {
-                  label: '&Reload',
-                  accelerator: 'Ctrl+R',
+                  label: "&Reload",
+                  accelerator: "Ctrl+R",
                   click: () => {
                     this.mainWindow.webContents.reload();
                   }
                 },
                 {
-                  label: 'Toggle &Full Screen',
-                  accelerator: 'F11',
+                  label: "Toggle &Full Screen",
+                  accelerator: "F11",
                   click: () => {
                     this.mainWindow.setFullScreen(
                       !this.mainWindow.isFullScreen()
@@ -228,8 +228,8 @@ export default class MenuBuilder {
                   }
                 },
                 {
-                  label: 'Toggle &Developer Tools',
-                  accelerator: 'Alt+Ctrl+I',
+                  label: "Toggle &Developer Tools",
+                  accelerator: "Alt+Ctrl+I",
                   click: () => {
                     this.mainWindow.webContents.toggleDevTools();
                   }
@@ -237,8 +237,8 @@ export default class MenuBuilder {
               ]
             : [
                 {
-                  label: 'Toggle &Full Screen',
-                  accelerator: 'F11',
+                  label: "Toggle &Full Screen",
+                  accelerator: "F11",
                   click: () => {
                     this.mainWindow.setFullScreen(
                       !this.mainWindow.isFullScreen()
@@ -248,32 +248,32 @@ export default class MenuBuilder {
               ]
       },
       {
-        label: 'Help',
+        label: "Help",
         submenu: [
           {
-            label: 'Learn More',
+            label: "Learn More",
             click() {
-              shell.openExternal('http://electron.atom.io');
+              shell.openExternal("http://electron.atom.io");
             }
           },
           {
-            label: 'Documentation',
+            label: "Documentation",
             click() {
               shell.openExternal(
-                'https://github.com/atom/electron/tree/master/docs#readme'
+                "https://github.com/atom/electron/tree/master/docs#readme"
               );
             }
           },
           {
-            label: 'Community Discussions',
+            label: "Community Discussions",
             click() {
-              shell.openExternal('https://discuss.atom.io/c/electron');
+              shell.openExternal("https://discuss.atom.io/c/electron");
             }
           },
           {
-            label: 'Search Issues',
+            label: "Search Issues",
             click() {
-              shell.openExternal('https://github.com/atom/electron/issues');
+              shell.openExternal("https://github.com/atom/electron/issues");
             }
           }
         ]
