@@ -2,9 +2,9 @@
  * Base webpack config used across other specific configs
  */
 
-import path from 'path';
-import webpack from 'webpack';
-import { dependencies } from '../package.json';
+import path from "path";
+import webpack from "webpack";
+import { dependencies } from "../package.json";
 
 export default {
   externals: [...Object.keys(dependencies || {})],
@@ -16,33 +16,33 @@ export default {
         exclude: /node_modules/,
         use: [
           {
-            loader: 'babel-loader',
+            loader: "babel-loader",
             options: {
               cacheDirectory: true
             }
           },
-          'ts-loader'
+          "ts-loader"
         ]
       }
     ]
   },
 
   output: {
-    path: path.join(__dirname, '..', 'app'),
+    path: path.join(__dirname, "..", "app"),
     // https://github.com/webpack/webpack/issues/1114
-    libraryTarget: 'commonjs2'
+    libraryTarget: "commonjs2"
   },
 
   /**
    * Determine the array of extensions that should be used to resolve modules.
    */
   resolve: {
-    extensions: ['.js', '.ts', '.tsx', '.json']
+    extensions: [".js", ".ts", ".tsx", ".json"]
   },
 
   plugins: [
     new webpack.EnvironmentPlugin({
-      NODE_ENV: 'production'
+      NODE_ENV: "production"
     }),
 
     new webpack.NamedModulesPlugin()
