@@ -1,4 +1,4 @@
 /* eslint import/no-unresolved: off, import/no-self-import: off */
-require('@babel/register');
+require("@babel/register");
 
-module.exports = require('./webpack.config.renderer.dev.babel').default;
+module.exports = require("./webpack.config.renderer.dev.babel").default;
