@@ -2,28 +2,28 @@
  * Webpack config for production electron main process
  */
 
-import path from 'path';
-import webpack from 'webpack';
-import merge from 'webpack-merge';
-import TerserPlugin from 'terser-webpack-plugin';
-import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer';
-import baseConfig from './webpack.config.base';
-import CheckNodeEnv from '../internals/scripts/CheckNodeEnv';
+import path from "path";
+import webpack from "webpack";
+import merge from "webpack-merge";
+import TerserPlugin from "terser-webpack-plugin";
+import { BundleAnalyzerPlugin } from "webpack-bundle-analyzer";
+import baseConfig from "./webpack.config.base";
+import CheckNodeEnv from "../internals/scripts/CheckNodeEnv";
 
-CheckNodeEnv('production');
+CheckNodeEnv("production");
 
 export default merge.smart(baseConfig, {
-  devtool: 'source-map',
+  devtool: "source-map",
 
-  mode: 'production',
+  mode: "production",
 
-  target: 'electron-main',
+  target: "electron-main",
 
-  entry: './app/main.dev',
+  entry: "./app/main.dev",
 
   output: {
-    path: path.join(__dirname, '..'),
-    filename: './app/main.prod.js'
+    path: path.join(__dirname, ".."),
+    filename: "./app/main.prod.js"
   },
 
   optimization: {
@@ -41,8 +41,8 @@ export default merge.smart(baseConfig, {
   plugins: [
     new BundleAnalyzerPlugin({
       analyzerMode:
-        process.env.OPEN_ANALYZER === 'true' ? 'server' : 'disabled',
-      openAnalyzer: process.env.OPEN_ANALYZER === 'true'
+        process.env.OPEN_ANALYZER === "true" ? "server" : "disabled",
+      openAnalyzer: process.env.OPEN_ANALYZER === "true"
     }),
 
     /**
@@ -55,7 +55,7 @@ export default merge.smart(baseConfig, {
      * development checks
      */
     new webpack.EnvironmentPlugin({
-      NODE_ENV: 'production',
+      NODE_ENV: "production",
       DEBUG_PROD: false,
       START_MINIMIZED: false
     })
