@@ -1,17 +1,17 @@
 // Check if the renderer and main bundles are built
-import path from 'path';
-import chalk from 'chalk';
-import fs from 'fs';
+import path from "path";
+import chalk from "chalk";
+import fs from "fs";
 
 function CheckBuildsExist() {
-  const mainPath = path.join(__dirname, '..', '..', 'app', 'main.prod.js');
+  const mainPath = path.join(__dirname, "..", "..", "app", "main.prod.js");
   const rendererPath = path.join(
     __dirname,
-    '..',
-    '..',
-    'app',
-    'dist',
-    'renderer.prod.js'
+    "..",
+    "..",
+    "app",
+    "dist",
+    "renderer.prod.js"
   );
 
   if (!fs.existsSync(mainPath)) {
