@@ -1,4 +1,4 @@
 /* eslint import/prefer-default-export: off */
-import { ClientFunction } from 'testcafe';
+import { ClientFunction } from "testcafe";
 
 export const getPageUrl = ClientFunction(() => window.location.href);
