@@ -1,5 +1,5 @@
-describe('description', () => {
-  it('should have description', () => {
+describe("description", () => {
+  it("should have description", () => {
     expect(1 + 2).toBe(3);
   });
 });
